requirements.txt
requests
pandas
streamlit
plotly
numpy
python-dotenv
google-generativeai
google-genai
fastapi
uvicorn
pydantic
